CSRNet-pytorch/requirements.txt
flask
flask-socketio
eventlet
boto3
livekit-api
python-dotenv
opencv-python-headless
numpy
torch
torchvision
werkzeug
mnest
python-engineio
python-socketio
opencv-python